requirements.txt
requests>=2.32.3
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
tenacity>=9.0.0
urllib3<2
fastapi>=0.115.0
uvicorn>=0.30.6
jinja2>=3.1.4
pypdf>=5.0.0
